fix: use layout instead of .columns
Why

- The .columns syntax is currently (quartro 1.3.450) sensitive to the
position of the tags in the html code. A better alternative is to use
the layout syntax that is not sensitive to linting change in the html
code. Those small changes are introduced when reading each html file
with rvest::read_html() that automatically escapes a line after each
ending tag.

How

- Modify the test qmd slides to use layout instead of .columns
- Remove post processing of html (removing line breack introduce by
  read_html())
- Modify the tests snapshot accordingly

Issue thinkr/thinkrverse/formation#222